MB-14732 rename files and update imports of component
