+++ src/components/LocationSearchBox/LocationSearchBox.test.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { render, screen, waitFor } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 
-import DutyLocationSearchBox from './DutyLocationSearchBox';
+import DutyLocationSearchBox from './LocationSearchBox';
 
 import dutyLocationFactory from 'utils/test/factories/dutyLocation';
